+++ ASP.NET-MVC/BookShop/BookShop.AcceptanceTests/Drivers/RowObjects/BookRow.cs
@@ -1,13 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace BookShop.AcceptanceTests.Drivers.RowObjects
+﻿namespace BookShop.AcceptanceTests.Drivers.RowObjects
 {
     class BookRow
     {
         public string? Title { get; set; }
         public string? Author { get; set; }
-        public float? Price { get; set; }
+        public decimal? Price { get; set; }
+        public string? Id { get; set; }
     }
 }
