+++ ASP.NET-MVC/BookShop/BookShop.AcceptanceTests/Drivers/Selenium/BrowserSeleniumDriverFactory.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.ObjectModel;
-using System.IO;
 using OpenQA.Selenium;
 using OpenQA.Selenium.Chrome;
 using OpenQA.Selenium.Edge;
