+++ ASP.NET-MVC/BookShop/BookShop.AcceptanceTests/Features/Home Screen.feature
@@ -7,18 +7,18 @@ Feature: Displaying Home Screen
 
 Background:
 	Given the following books
-		|Author			|Title								|Price	|
-		|Martin Fowler	|Analysis Patterns					|50.20	|
-		|Eric Evans		|Domain Driven Design				|46.34	|
-		|Ted Pattison	|Inside Windows SharePoint Services	|31.49	|
-		|Gojko Adzic	|Bridging the Communication Gap		|24.75	|
+		| Title                              | Price |
+		| Analysis Patterns                  | 50.20 |
+		| Domain Driven Design               | 46.34 |
+		| Inside Windows SharePoint Services | 31.49 |
+		| Bridging the Communication Gap     | 24.75 |
 
 @WI8
 Scenario: Cheapest 3 books should be listed on the home screen
 	When I enter the shop
 	Then the home screen should show the book 'Bridging the Communication Gap'
-	   * the home screen should show the book 'Inside Windows SharePoint Services'
-	   * the home screen should show the book 'Domain Driven Design'
+	And the home screen should show the book 'Inside Windows SharePoint Services'
+	And the home screen should show the book 'Domain Driven Design'
 
 @alternative_syntax
 @WI8
@@ -31,9 +31,7 @@ Scenario: Cheapest 3 books should be listed on the home screen (list syntax)
 Scenario: Cheapest 3 books should be listed on the home screen (table syntax)
 	When I enter the shop
 	Then the home screen should show the following books
-		|Title								|
-		|Bridging the Communication Gap		|
-		|Inside Windows SharePoint Services	|
-		|Domain Driven Design				|
-
-
+		| Title                              |
+		| Bridging the Communication Gap     |
+		| Inside Windows SharePoint Services |
+		| Domain Driven Design               |
