+++ ASP.NET-MVC/BookShop/BookShop.AcceptanceTests/Features/Shopping Cart/Display.feature
@@ -7,11 +7,11 @@ Feature: Displaying the shopping cart
 
 Background:
 	Given the following books
-		|Author			|Title								|Price	|
-		|Martin Fowler	|Analysis Patterns					|50.20	|
-		|Eric Evans		|Domain Driven Design				|46.34	|
-		|Ted Pattison	|Inside Windows SharePoint Services	|31.49	|
-		|Gojko Adzic	|Bridging the Communication Gap		|24.75	|
+		| Title                              | Price |
+		| Analysis Patterns                  | 50.20 |
+		| Domain Driven Design               | 46.34 |
+		| Inside Windows SharePoint Services | 31.49 |
+		| Bridging the Communication Gap     | 24.75 |
 
 @WI15
 Scenario: Shopping cart should show total number of items and total price
@@ -21,8 +21,6 @@ Scenario: Shopping cart should show total number of items and total price
 	And my shopping cart should contain 3 items in total
 	And my shopping cart should show a total price of 146.74
 
-
 Scenario: The shopping cart should be initially empty
 	When I enter the shop
 	Then my shopping cart should be empty
-
